Refactor stream module into submodules and fix tests
